Make Savon CVE fix surgical
