+++ lib/active_zuora/savon_model_cve_2026_53510_patch.rb
@@ -1,0 +1,19 @@
+module ActiveZuora
+  module SavonModelCve202653510Patch
+    def actions(*actions)
+      actions.each do |action|
+        method_name = action.to_s.snakecase.to_sym
+
+        class_action_module.define_method(method_name) do |body = nil, &block|
+          client.request :wsdl, action, :body => body, &block
+        end
+
+        instance_action_module.define_method(method_name) do |body = nil, &block|
+          self.class.public_send(method_name, body, &block)
+        end
+      end
+    end
+  end
+end
+
+Savon::Model.prepend(ActiveZuora::SavonModelCve202653510Patch)
