--- lib/savon/error.rb
@@ -1,6 +1,0 @@
-module Savon
-
-  # Base class for Savon errors.
-  class Error < RuntimeError; end
-
-end
