--- lib/savon/null_logger.rb
@@ -1,10 +1,0 @@
-require "savon/logger"
-
-module Savon
-  class NullLogger < Logger
-
-    def log(*)
-    end
-
-  end
-end
