--- lib/savon/soap/invalid_response_error.rb
@@ -1,13 +1,0 @@
-require "savon/error"
-
-module Savon
-  module SOAP
-
-    # = Savon::SOAP::InvalidResponseError
-    #
-    # Represents an error when the response was not a valid SOAP envelope.
-    class InvalidResponseError < Error
-    end
-
-  end
-end
