--- lib/savon/version.rb
@@ -1,5 +1,0 @@
-module Savon
-
-  VERSION = "1.2.0"
-
-end
